# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/kakao": {
      "post": {
        "operationId": "AuthController_kakao",
        "summary": "카카오 로그인/가입 (화이트리스트 검증)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KakaoLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "리프레시 토큰 회전",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "로그아웃 (리프레시 토큰 무효화)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "로그아웃 완료(본문 없음)"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "summary": "내 정보 + 온보딩 상태",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/media/upload-url": {
      "post": {
        "operationId": "MediaController_createUploadUrl",
        "summary": "업로드용 presigned URL 발급",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadUrlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadUrlResponse"
                }
              }
            }
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/media/{mediaId}/complete": {
      "post": {
        "operationId": "MediaController_complete",
        "summary": "업로드 완료 확정 (객체 존재 확인 + READY)",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            }
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/media/{mediaId}": {
      "get": {
        "operationId": "MediaController_getMedia",
        "summary": "미디어 메타데이터 + 다운로드 URL (권한 검증)",
        "parameters": [
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponse"
                }
              }
            }
          }
        },
        "tags": [
          "media"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/rooms/{roomId}": {
      "get": {
        "operationId": "RoomsController_detail",
        "summary": "방 상세 + 멤버 목록",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomViewResponse"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "summary": "방 정보 수정 (그룹, OWNER/ADMIN)",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomViewResponse"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/rooms/{roomId}/messages": {
      "get": {
        "operationId": "MessagesController_list",
        "summary": "메시지 이력 (커서 페이지네이션, seq < cursor 인 과거 메시지)",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "커서(seq 또는 id 또는 ISO 타임스탬프)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponse"
                }
              }
            }
          }
        },
        "tags": [
          "messages"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "MessagesController_send",
        "summary": "메시지 전송 (REST 폴백). 기본 전송은 WebSocket 사용",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMessageResponse"
                }
              }
            }
          }
        },
        "tags": [
          "messages"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rooms/{roomId}/messages/{messageId}": {
      "delete": {
        "operationId": "MessagesController_remove",
        "summary": "메시지 삭제(언센드) — 발신자 본인",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "언센드 완료(본문 없음)"
          }
        },
        "tags": [
          "messages"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rooms/{roomId}/read": {
      "post": {
        "operationId": "MessagesController_read",
        "summary": "읽음 위치 갱신",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "읽음 위치 갱신 완료(본문 없음)"
          }
        },
        "tags": [
          "messages"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/whitelist": {
      "post": {
        "operationId": "AdminController_create",
        "summary": "화이트리스트 등록/초대",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWhitelistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhitelistResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "AdminController_list",
        "summary": "화이트리스트 목록",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "커서(seq 또는 id 또는 ISO 타임스탬프)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "INVITED",
                "ACTIVE",
                "BLOCKED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedWhitelistResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/whitelist/{id}": {
      "patch": {
        "operationId": "AdminController_update",
        "summary": "화이트리스트 상태 변경 (BLOCKED 시 세션 무효화)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWhitelistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhitelistResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_remove",
        "summary": "화이트리스트 삭제",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 완료(본문 없음)"
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "인택톡(Intaktok) API",
    "description": "폐쇄형 실시간 채팅 플랫폼 백엔드 API (v2.0, E2EE 제외)",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "DeviceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "example": "iPhone 15"
          },
          "platform": {
            "type": "string",
            "enum": [
              "ios",
              "android",
              "web"
            ]
          }
        }
      },
      "KakaoLoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "카카오 로그인 리다이렉트 콜백으로 받은 인가 코드(authorization code). 백엔드가 앱 시크릿과 함께 카카오 서버에 제출해 액세스 토큰으로 교환합니다."
          },
          "redirectUri": {
            "type": "string",
            "description": "인가 코드를 발급받을 때 사용한 redirect_uri. 카카오 토큰 교환 시 바이트 단위로 동일해야 합니다.",
            "example": "https://intaektalk.app/oauth/kakao/callback"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceDto"
          }
        },
        "required": [
          "code",
          "redirectUri"
        ]
      },
...
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "액세스 토큰(JWT)"
          },
          "refreshToken": {
            "type": "string",
            "description": "리프레시 토큰(JWT, 회전 대상)"
          },
          "user": {
            "$ref": "#/components/schemas/LoginUserResponse"
          },
          "isOnboarded": {
            "type": "boolean",
            "description": "온보딩(프로필 생성) 완료 여부"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user",
          "isOnboarded"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TokenPairResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "액세스 토큰(JWT)"
          },
          "refreshToken": {
            "type": "string",
            "description": "리프레시 토큰(JWT, 회전 대상)"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "MeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kakaoId": {
            "type": "string",
            "description": "카카오 회원번호"
          },
          "isOnboarded": {
            "type": "boolean"
          },
          "profile": {
            "nullable": true,
            "description": "온보딩 전이면 null",
            "allOf": [
              {
                "$ref": "#/components/schemas/MeProfileResponse"
              }
            ]
          }
        },
        "required": [
          "id",
          "kakaoId",
          "isOnboarded",
          "profile"
        ]
      },
      "CreateUploadUrlDto": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "pattern": "/^[\\w.+-]+\\/[\\w.+-]+$/",
            "example": "image/gif"
          },
          "byteSize": {
            "type": "number",
            "minimum": 1,
            "maximum": 1073741824,
            "example": 1048576,
            "description": "바이트 크기"
          },
          "width": {
            "type": "number",
            "minimum": 1
          },
          "height": {
            "type": "number",
            "minimum": 1
          },
          "durationMs": {
            "type": "number",
            "minimum": 0,
            "description": "동영상/애니메이션 길이(ms)"
          }
        },
        "required": [
          "mimeType",
          "byteSize"
        ]
      },
      "UploadUrlResponse": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string",
            "description": "업로드용 presigned PUT URL"
          },
          "storageKey": {
            "type": "string",
            "description": "스토리지 객체 키"
          }
        },
        "required": [
          "mediaId",
          "uploadUrl",
          "storageKey"
        ]
      },
      "MediaResponse": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "description": "다운로드(GET) URL"
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true,
            "description": "썸네일 URL(없으면 null)"
          },
          "mimeType": {
            "type": "string",
            "example": "image/gif"
          },
          "isAnimated": {
            "type": "boolean",
            "description": "자동재생 대상 여부(움짤/영상)"
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "durationMs": {
            "type": "number",
            "nullable": true,
            "description": "동영상/애니메이션 길이(ms)"
          }
        },
        "required": [
          "mediaId",
          "url",
          "thumbnailUrl",
          "mimeType",
          "isAnimated",
          "width",
          "height",
          "durationMs"
        ]
      },
...
      "RoomViewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DIRECT",
              "GROUP"
            ]
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "GROUP 방 이름"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastMessageAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomMemberResponse"
            }
          },
          "lastMessage": {
            "nullable": true,
            "description": "마지막 메시지(없으면 null)",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomLastMessageResponse"
              }
            ]
          },
          "unreadCount": {
            "type": "number",
            "description": "미읽음 메시지 수"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "avatarUrl",
          "createdBy",
          "createdAt",
          "lastMessageAt",
          "members"
        ]
      },
...
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "avatarMediaId": {
            "type": "string"
          }
        }
      },
...
      "PaginatedMessagesResponse": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "다음 페이지 커서. null이면 마지막 페이지입니다."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponse"
            }
          }
        },
        "required": [
          "nextCursor",
          "items"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "clientMessageId": {
            "type": "string",
            "maxLength": 64,
            "description": "멱등성 키(클라이언트 UUID)"
          },
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "IMAGE",
              "GIF",
              "VIDEO",
              "FILE"
            ]
          },
          "content": {
            "type": "string",
            "nullable": true,
            "maxLength": 8000,
            "description": "TEXT 본문 또는 미디어 캡션"
          },
          "mediaId": {
            "type": "string",
            "nullable": true,
            "description": "미디어 메시지일 때 mediaId"
          },
          "replyToId": {
            "type": "string",
            "nullable": true,
            "description": "답장 대상 메시지 id"
          }
        },
        "required": [
          "type"
        ]
      },
      "SendMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "seq": {
            "type": "number",
            "description": "발급된 시퀀스"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "seq",
          "createdAt"
        ]
      },
      "ReadDto": {
        "type": "object",
        "properties": {
          "lastReadSeq": {
            "type": "number",
            "minimum": 0,
            "example": 42
          }
        },
        "required": [
          "lastReadSeq"
        ]
      },
...
      "CreateWhitelistDto": {
        "type": "object",
        "properties": {
          "kakaoId": {
            "type": "string",
            "description": "카카오 회원번호"
          },
          "identifier": {
            "type": "string",
            "maxLength": 200,
            "description": "대체 식별자(이메일/전화 등)"
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "WhitelistResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kakaoId": {
            "type": "string",
            "nullable": true,
            "description": "카카오 회원번호"
          },
          "identifier": {
            "type": "string",
            "nullable": true,
            "description": "대체 식별자(이메일/전화 등)"
          },
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "BLOCKED"
            ]
          },
          "invitedBy": {
            "type": "string",
            "nullable": true,
            "description": "초대한 사용자 id"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "kakaoId",
          "identifier",
          "status",
          "invitedBy",
          "note",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedWhitelistResponse": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "다음 페이지 커서. null이면 마지막 페이지입니다."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WhitelistResponse"
            }
          }
        },
        "required": [
          "nextCursor",
          "items"
        ]
      },
      "UpdateWhitelistDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "INVITED",
              "ACTIVE",
              "BLOCKED"
            ]
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
    }
  }
}